+++ CHANGELOG.md
@@ -59,9 +59,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* [#753](https://github.com/allora-network/allora-chain/pull/753) Add GRPC support to the testing localnet
+
 ### Changed
 
 * [#762](https://github.com/allora-network/allora-chain/pull/762) Worker node owner should get compensated, not sender of inferences
+* [#766](https://github.com/allora-network/allora-chain/pull/766) Fix sort by weight instead of topic id. Remove excess sort logic
 
 ### Deprecated
 
@@ -70,8 +73,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * [#765](https://github.com/allora-network/allora-chain/pull/765) Fix rounding precision issues in mint module
+* [#757](https://github.com/allora-network/allora-chain/pull/757) Incoming value limiting
 
 ### Security
+
+* [#756](https://github.com/allora-network/allora-chain/pull/756) Prevent ecosystem rewards after its supply hit
+* [#767](https://github.com/allora-network/allora-chain/pull/767) Fix [ISA-2025-001](https://github.com/cosmos/ibc-go/security/advisories/GHSA-4wf3-5qj9-368v)
+
+# [Released]
 
 # v0.9.0
 
@@ -173,8 +182,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#713](https://github.com/allora-network/allora-chain/pull/713) Proto generation keeps codec.go files
 
 ### Security
-
-# [Released]
 
 ## v0.7.0